Release pré-apresentação
* Removidas entidades desnecessárias
* Simplificada estrutura dos Wishlist Items
* Ordenados itens por prioridade
