--- src/main/java/com/kappa/kindly/repository/DonationItemRepository.java
@@ -1,13 +1,0 @@
-package com.kappa.kindly.repository;
-
-import com.kappa.kindly.model.DonationItem;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface DonationItemRepository extends JpaRepository<DonationItem, Long> {
-	
-	DonationItem findById(long id);
-
-}
